refactor: split 3 files that crossed the 800-line file-shape cap (#1514)
frontend/src/offline/media.rs and frontend/src/offline/downloads.rs
each crossed the ~800-line soft cap; split by sub-topic, mirroring
the existing db/src/indexer.rs and offline/downloads/{engine,verify}
precedent:

- offline/media.rs -> media/revalidate.rs: the background ETag
  revalidation machinery (revalidate_in_background, claim/release,
  lock_cache_key, the Fetched enum).
- offline/downloads.rs -> downloads/staleness.rs: "Update available"
  detection and its two remedies (is_stale/note_book,
  refresh_stale_flags's TTL-gated validators poll, and redownload).

server/src/backend.rs (~801 lines per the issue) is already at 785
lines on this branch's base — a prior refactor (#1520) deduplicated
its 500-response helper and brought it back under the cap on its
own, so it needs no further split here.

Pure code motion: same functions, same logic, only file/module
location and the pub/pub(super)/pub(crate) visibility needed to
preserve every external call path changed.
